# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GreenLogix API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/optimize": {
      "post": {
        "tags": [
          "optimize"
        ],
        "summary": "Optimize",
        "operationId": "optimize_optimize_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routes": {
      "get": {
        "tags": [
          "optimize"
        ],
        "summary": "List Routes",
        "operationId": "list_routes_routes_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteOut"
                  },
                  "title": "Response List Routes Routes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routes/publish": {
      "post": {
        "tags": [
          "optimize"
        ],
        "summary": "Publish Routes",
        "operationId": "publish_routes_routes_publish_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteOut"
                  },
                  "title": "Response Publish Routes Routes Publish Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routes/{id}/admin-areas": {
      "get": {
        "tags": [
          "optimize"
        ],
        "summary": "Route Admin Areas",
        "description": "Versioned admin-area corridor for one route (T-LOOP-GEO, CR-08).\n\nCentroid fallback labeled ``nso-2024-v1-centroid-fallback``; never claims\npolygon intersection. Root-mounted prefix matches existing ``/routes``\nstyle (no ``/v1`` prefix anywhere in this API).",
        "operationId": "route_admin_areas_routes__id__admin_areas_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Route Admin Areas Routes  Id  Admin Areas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/report": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Report",
        "operationId": "report_report_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/report.xlsx": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Report Xlsx",
        "operationId": "report_xlsx_report_xlsx_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptimizeIn": {
        "properties": {
          "cluster_radius_km": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cluster Radius Km",
            "default": 3.0
          }
        },
        "type": "object",
        "title": "OptimizeIn"
      },
      "OptimizeOut": {
        "properties": {
          "routes": {
            "items": {
              "$ref": "#/components/schemas/RouteOut"
            },
            "type": "array",
            "title": "Routes"
          },
          "unassigned_order_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Unassigned Order Ids"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsOut"
          },
          "baseline": {
            "$ref": "#/components/schemas/TotalsOut"
          },
          "distance_provider": {
            "type": "string",
            "title": "Distance Provider",
            "default": "circuity"
          },
          "routing_quality": {
            "type": "string",
            "title": "Routing Quality",
            "default": "DEGRADED"
          },
          "eco_weight": {
            "type": "number",
            "title": "Eco Weight",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "routes",
          "unassigned_order_ids",
          "totals",
          "baseline"
        ],
        "title": "OptimizeOut"
      },
...
      "PublishIn": {
        "properties": {
          "route_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Route Ids"
          }
        },
        "type": "object",
        "title": "PublishIn"
      },
      "ReportDelta": {
        "properties": {
          "km": {
            "type": "number",
            "title": "Km"
          },
          "litres": {
            "type": "number",
            "title": "Litres"
          },
          "kg_co2": {
            "type": "number",
            "title": "Kg Co2"
          },
          "km_pct": {
            "type": "number",
            "title": "Km Pct"
          },
          "litres_pct": {
            "type": "number",
            "title": "Litres Pct"
          },
          "kg_co2_pct": {
            "type": "number",
            "title": "Kg Co2 Pct"
          }
        },
        "type": "object",
        "required": [
          "km",
          "litres",
          "kg_co2",
          "km_pct",
          "litres_pct",
          "kg_co2_pct"
        ],
        "title": "ReportDelta"
      },
      "ReportOut": {
        "properties": {
          "baseline": {
            "$ref": "#/components/schemas/ReportTotals"
          },
          "optimized": {
            "$ref": "#/components/schemas/ReportTotals"
          },
          "delta": {
            "$ref": "#/components/schemas/ReportDelta"
          },
          "distance_provider": {
            "type": "string",
            "title": "Distance Provider",
            "default": "circuity"
          },
          "routing_quality": {
            "type": "string",
            "title": "Routing Quality",
            "default": "DEGRADED"
          },
          "eco_weight": {
            "type": "number",
            "title": "Eco Weight",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "baseline",
          "optimized",
          "delta"
        ],
        "title": "ReportOut"
      },
      "ReportTotals": {
        "properties": {
          "km": {
            "type": "number",
            "title": "Km"
          },
          "litres": {
            "type": "number",
            "title": "Litres"
          },
          "kg_co2": {
            "type": "number",
            "title": "Kg Co2"
          }
        },
        "type": "object",
        "required": [
          "km",
          "litres",
          "kg_co2"
        ],
        "title": "ReportTotals"
      },
      "RouteOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "vehicle_id": {
            "type": "integer",
            "title": "Vehicle Id"
          },
          "plate": {
            "type": "string",
            "title": "Plate"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "published": {
            "type": "boolean",
            "title": "Published"
          },
          "km": {
            "type": "number",
            "title": "Km"
          },
          "litres": {
            "type": "number",
            "title": "Litres"
          },
          "kg_co2": {
            "type": "number",
            "title": "Kg Co2"
          },
          "overload": {
            "type": "boolean",
            "title": "Overload"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/StopOut"
            },
            "type": "array",
            "title": "Stops"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle_id",
          "plate",
          "color",
          "published",
          "km",
          "litres",
          "kg_co2",
          "overload",
          "stops"
        ],
        "title": "RouteOut"
      },
...
      "StopOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "kind": {
            "type": "string",
            "enum": [
              "depot",
              "stop"
            ],
            "title": "Kind"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "window_start": {
            "type": "string",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "title": "Window End"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "kg": {
            "type": "number",
            "title": "Kg"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "arrived",
              "delivered",
              "failed"
            ],
            "title": "Status"
          },
          "fail_reason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "khach_vang",
                  "sai_dia_chi",
                  "hang_hong",
                  "tu_choi"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Fail Reason"
          },
          "late_risk": {
            "type": "boolean",
            "title": "Late Risk",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "seq",
          "kind",
          "order_id",
          "lat",
          "lng",
          "address",
          "phone",
          "window_start",
          "window_end",
          "notes",
          "kg",
          "status"
        ],
        "title": "StopOut"
      },
      "TotalsOut": {
        "properties": {
          "km": {
            "type": "number",
            "title": "Km"
          },
          "litres": {
            "type": "number",
            "title": "Litres"
          },
          "kg_co2": {
            "type": "number",
            "title": "Kg Co2"
          }
        },
        "type": "object",
        "required": [
          "km",
          "litres",
          "kg_co2"
        ],
        "title": "TotalsOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}